# large txn performance improvements