--- a/dev/customers/icebreakers.clj
+++ b/dev/customers/icebreakers.clj
@@ -1,1 +1,0 @@
-(ns icebreakers)