This gives a significant boost in transaction times, although significant increases are most noticable with larger transactions.

This improved transaction time by > 30% in a transaction containing 1 million flakes / 15MB of data.

Txn time went from ~28 seconds to ~19 seconds.

The improvements are achieved by running some operations in parallel on the JVM that were coded as serial.

Notable:
1) Calculating flake-size (for db size) in commit data is much slower than I expected. Not sure what the bottleneck is, didn't investigate - but of original 28 seconds, this operation alone took 6 seconds. This is done in a background thread now.
2) post, opst updates takes about 4x longer than spot, tspo indexes. I think this makes sense because there are large numbers of the same 'p' in most data sets, so the first equivalency tests still will have a large number of secondary items to check. These two indexes only are updated in the background... spot & tspo will both finish before post, so left those as-is for now.